"EAP-related" code has been put in a separate module.
